Extract GAPI request example into separate file
